requirements.txt
keboola.component==1.3.5
keboola.http-client
mock
freezegun
requests==2.28.1
keboola.csvwriter==1.0.1
